Move awaitable CallResult to a seperate file in Samples folder.
This commit changed the way how Unity user consume this feature. I imagine that Unity consumers will use this feature by copy Samples\CallResultAwaitable.cs into their project. For .NET Standard users, they will have this as built-in feature, since Awaitable CallResult used internal API(CallbackDispatcher class) that can't accessed outside our DLL.
